mod: 핸들러, 함수 분리 작업..
