+++ index.html
@@ -34,6 +34,8 @@
         </div>
       </div>
     </div>
-    <script src="src/js/index.js"></script>
+    <script type="module" src="src/js/handler.js"></script>
+    <script type="module" src="src/js/function/arithmetic.js"></script>
+    <script type="module" src="src/js/index.js"></script>
   </body>
 </html>
